Use wallet proxy to get terms and conditions version
