+++ packages/browser-wallet/src/popup/shell/Root.tsx
@@ -13,7 +13,7 @@ import { absoluteRoutes } from '@popup/constants/routes';
 
 import './i18n';
 
-import { mainnet } from '@popup/pages/NetworkSettings/NetworkSettings';
+import { mainnet } from '@shared/constants/networkConfiguration';
 import Routes from './Routes';
 
 const body = document.getElementsByTagName('body').item(0);
