+++ packages/browser-wallet/src/shared/utils/network-helpers.ts
@@ -1,11 +1,9 @@
-import { Network, JsonRpcClient, CryptographicParameters, sha256 } from '@concordium/web-sdk';
+import { Network, JsonRpcClient, CryptographicParameters } from '@concordium/web-sdk';
 import { NetworkConfiguration } from '@shared/storage/types';
-import urls from '@shared/constants/url';
-
-export const mainnetGenesisHash = '9dd9ca4d19e9393877d2c44b70f89acbfc0883c2243e5eeaecc0d1cd0503f478';
+import { mainnet } from '@shared/constants/networkConfiguration';
 
 export function isMainnet(network: NetworkConfiguration) {
-    return network.genesisHash === mainnetGenesisHash;
+    return network.genesisHash === mainnet.genesisHash;
 }
 
 export function getNet(network: NetworkConfiguration): Network {
@@ -20,7 +18,9 @@ export async function getGlobal(client: JsonRpcClient): Promise<CryptographicPar
     return global.value;
 }
 
-export async function getTermsAndConditionHash() {
-    const site = await (await fetch(urls.termsAndConditions)).arrayBuffer();
-    return sha256([new Uint8Array(site)]).toString('base64');
+export async function getTermsAndConditionsConfig(
+    explorerUrl: string = mainnet.explorerUrl
+): Promise<{ version: string; url: string } | undefined> {
+    const response = await fetch(`${explorerUrl}/v0/termsAndConditionsVersion`);
+    return response.json();
 }
